Changed Orders to decimal and numeric

+++ db/migration/000005_add_order_table.up.sql
@@ -2,9 +2,9 @@ CREATE TABLE "orders" (
   "id" bigserial PRIMARY KEY,
   "owner" varchar NOT NULL,
   "status" varchar NOT NULL,
-  "delivery_fee" bigint NOT NULL,
-  "subtotal" bigint NOT NULL,
-  "total" bigint NOT NULL,
+  "delivery_fee" numeric NOT NULL,
+  "subtotal" numeric NOT NULL,
+  "total" numeric NOT NULL,
   "created_at" timestamptz NOT NULL DEFAULT 'now()'
 );
 ALTER TABLE "orders" ADD FOREIGN KEY ("owner") REFERENCES "users" ("username");
